update code to filter for new treatments and refactor shap waterfall plot
