+++ src/dashboard/generate_dashboard_per_patient.py
@@ -3,7 +3,6 @@ Module to assemble and save patient dashboards as PNGs using Playwright.
 """
 
 import base64
-import os
 from pathlib import Path
 
 import pandas as pd
@@ -12,9 +11,10 @@ from playwright.sync_api import sync_playwright
 
 from deployer.dashboard.component import create_patient_overview, create_model_overview, create_percentile_overview
 from deployer.dashboard.risk_dist_plot import risk_dist_plot
+from deployer.dashboard.shap_waterfall import SHAP_SUBDIR, build_shap_explainer, save_shap_waterfall_plot
+from deployer.loader import Model
 
 DASHBOARD_SUBDIR = "dashboards"
-SHAP_SUBDIR = "shap_waterfall"
 SUPPORTED_LAYOUTS = {"portrait", "landscape"}
 DEFAULT_FONT_SCALE = 1.0
 
@@ -111,6 +111,8 @@ def _build_dashboard_html(
 
 
 def save_dashboard_png(
+    model: Model,
+    df_input: pd.DataFrame,
     df_output: pd.DataFrame,
     df_meta: pd.DataFrame,
     output_dir: str | Path,
@@ -124,6 +126,8 @@ def save_dashboard_png(
     The SHAP waterfall PNG is deleted after use.
 
     Args:
+        model: Trained model bundle used to compute SHAP waterfall plots.
+        df_input: DataFrame of model inputs aligned 1:1 with df_output.
         df_output: DataFrame with columns: mrn, next_sched_trt_date, clinic_date,
                    ed_pred_prob, ed_pred_alarm_0.1
         df_meta:   DataFrame with columns: mrn, clinic_date, age, gender, cancer
@@ -149,7 +153,18 @@ def save_dashboard_png(
             f"Silent deployment baseline file not found: {baseline_path}"
         )
 
+    if len(df_input) != len(df_output):
+        raise ValueError("df_input and df_output must have the same number of rows")
+
+    if df_output.empty:
+        return
+
     df_baseline = pd.read_csv(baseline_path, usecols=["mrn", "ed_pred_prob", "cancer"])
+    shap_dir.mkdir(parents=True, exist_ok=True)
+    explainer = build_shap_explainer(model)
+
+    df_meta = df_meta.copy()
+    df_meta["clinic_date"] = pd.to_datetime(df_meta["clinic_date"])
 
     # Index df_meta by (mrn, clinic_date) for fast lookup
     meta_indexed = df_meta.set_index(["mrn", "clinic_date"])
@@ -159,18 +174,20 @@ def save_dashboard_png(
         browser = p.chromium.launch()
         page = browser.new_page(viewport=viewport)
 
-        df_meta["clinic_date"] = pd.to_datetime(df_meta["clinic_date"])
-
-        for _, row in df_output.iterrows():
+        for row_idx, row in df_output.reset_index(drop=True).iterrows():
             mrn = int(row["mrn"])
             clinic_date = pd.Timestamp(row["clinic_date"])
             clinic_date_str = clinic_date.strftime("%Y%m%d")
 
-            # Locate SHAP PNG
             shap_png_path = shap_dir / f"shap_mrn_{mrn}_clinic_{clinic_date_str}.png"
-            if not shap_png_path.exists():
-                print(f"[WARN] SHAP plot not found, skipping: {shap_png_path}")
-                continue
+            save_shap_waterfall_plot(
+                model=model,
+                model_input=df_input,
+                explainer=explainer,
+                output_path=shap_png_path,
+                row_idx=row_idx,
+                font_scale=font_scale,
+            )
 
             # Fetch matching meta row (most recent clinic_date <= current for this mrn)
             try:
