+++ src/main.py
@@ -3,10 +3,11 @@ import os
 import warnings
 
 import pandas as pd
+from deployer.dashboard.filter import get_dashboard_keep_mask
 from deployer.data_prep.pipeline import build_features, get_data
+from deployer.dashboard.generate_dashboard_per_patient import save_dashboard_png
 from deployer.loader import Config, Model
 from deployer.model_eval.inference import get_model_output
-from deployer.dashboard.generate_dashboard_per_patient import save_dashboard_png
 from tqdm import tqdm
 
 warnings.filterwarnings("ignore")
@@ -23,6 +24,12 @@ def parse_args():
         "--disable-save-dashboard-png",
         action="store_true",
         help="Skip generating dashboard PNG files.",
+    )
+    parser.add_argument(
+        "--subset-dashboard-patients",
+        action=argparse.BooleanOptionalAction,
+        default=True,
+        help="Generate dashboards only for the selected subset of patients. Default is True.",
     )
     parser.add_argument("--run-on-silent-deployment", type=bool, default=False)
 
@@ -42,6 +49,7 @@ if __name__ == "__main__":
     dashboard_layout = args.dashboard_layout
     dashboard_font_scale = args.dashboard_font_scale
     disable_save_dashboard_png = args.disable_save_dashboard_png
+    subset_dashboard_patients = args.subset_dashboard_patients
     run_on_silent_deployment = args.run_on_silent_deployment
     output_dir = args.output_dir
     data_dir = args.data_dir
@@ -60,7 +68,7 @@ if __name__ == "__main__":
     thresholds = config.thresholds.query(f'model_anchor == "{anchor.title()}-anchored"')
 
     date_range = pd.date_range(start_date, end_date, freq="d").strftime("%Y%m%d")
-    inputs, outputs, meta_data = [], [], []
+    inputs, outputs, meta_data, dashboard_masks = [], [], [], []
     for i, data_pull_date in tqdm(enumerate(date_range)):
         print(f"**** Processing #{i}: {data_pull_date} *****")
         feats = build_features(config, data_dir, data_pull_date, model.anchor)
@@ -77,24 +85,43 @@ if __name__ == "__main__":
             feats['demographic'],
             thresholds,
             pred_fn=None,
-            output_dir=output_dir,
-            dashboard_font_scale=dashboard_font_scale,
-            disable_save_dashboard_png=disable_save_dashboard_png
         )
 
         inputs.append(res["model_input"])
         outputs.append(res["model_output"])
         meta_data.append(res["demographic_info"])
+        if subset_dashboard_patients:
+            dashboard_masks.append(
+                get_dashboard_keep_mask(
+                    res["model_output"],
+                    data_dir,
+                    data_pull_date,
+                    model.anchor,
+                )
+            )
+        else:
+            dashboard_masks.append(pd.Series(1, index=res["model_output"].index, dtype=int))
 
     out = pd.concat(outputs, ignore_index=True, axis=0)
+    out = out.reset_index(drop=True)
 
     inp = pd.concat(inputs, ignore_index=True, axis=0)
+    inp = inp.reset_index(drop=True)
     inp.to_parquet(f"{output_dir}/input_{anchor}.parquet")
 
     meta = pd.concat(meta_data, ignore_index=True, axis=0)
+    meta = meta.reset_index(drop=True)
+
+    mask = pd.concat(dashboard_masks, ignore_index=True, axis=0)
+    mask = mask.reset_index(drop=True)
+
+    dashboard_inp = inp.loc[mask == 1].reset_index(drop=True)
+    dashboard_meta = meta.loc[mask == 1].reset_index(drop=True)
 
     out = out.merge(meta[['mrn', 'clinic_date', 'cancer']], on=['mrn', 'clinic_date'], how='left')
     out.to_csv(f"{output_dir}/output_{anchor}.csv", index=False)
+
+    dashboard_out = out.loc[mask == 1].reset_index(drop=True)
 
     if run_on_silent_deployment:
         # filter out patients from out
@@ -110,8 +137,10 @@ if __name__ == "__main__":
     # Generate dashboard per patient
     if not disable_save_dashboard_png:
         save_dashboard_png(
-            out,
-            meta,
+            model,
+            dashboard_inp,
+            dashboard_out,
+            dashboard_meta,
             output_dir,
             anchor=anchor,
             layout=dashboard_layout,
